# Generate links to definition in rustdoc source code pages