--- a/src/test/rustdoc-ui/generate-link-to-definition-opt-unstable.stderr
+++ b/src/test/rustdoc-ui/generate-link-to-definition-opt-unstable.stderr
@@ -1,0 +1,2 @@
+error: the `-Z unstable-options` flag must also be passed to enable the flag `generate-link-to-definition`
+