--- a/src/test/rustdoc/auxiliary/source_code.rs
+++ b/src/test/rustdoc/auxiliary/source_code.rs
@@ -1,0 +1,1 @@
+pub struct SourceCode;