… src/librustdoc/html/highlight/fixtures/highlight.html
… src/librustdoc/html/highlight/fixtures/sample.html
… src/librustdoc/html/highlight/tests.rs
… src/librustdoc/html/markdown.rs
… src/librustdoc/html/render/context.rs
… src/librustdoc/html/render/mod.rs
… src/librustdoc/html/render/print_item.rs
… src/librustdoc/html/render/span_map.rs
… src/librustdoc/html/render/write_shared.rs
… src/librustdoc/html/sources.rs
… src/librustdoc/html/static/css/rustdoc.css
… src/librustdoc/lib.rs
… src/test/rustdoc-ui/generate-link-to-definition-opt-unstable.rs
… src/test/rustdoc-ui/generate-link-to-definition-opt.rs
… src/test/rustdoc-ui/generate-link-to-definition-opt.stderr
… src/test/rustdoc-ui/generate-link-to-definition-opt2.rs
… src/test/rustdoc-ui/generate-link-to-definition-opt2.stderr
… src/test/rustdoc/auxiliary/source-code-bar.rs
… src/test/rustdoc/check-source-code-urls-to-def.rs